+++ src/components/ImageHandlerContext/ImageHandlerContext.tsx
@@ -15,6 +15,7 @@ const ImageHandlerContext: React.FunctionComponent<
     useHttps: true,
     srcSetSizes: [480, 768, 992, 1280, 1920, 2048, 3840],
     optimizeSvg: false,
+    optimizeGif: false,
     progressiveImageLoading: true,
     // User's configuration
     ...config,